[LABS-477] Basedpyright fixes for wallet tests (#21061)
* [LABS-470] Fix mismatching Wallets with WalletProtocol

* [LABS-471] Remove contravariant from WalletProtocol

* [LABS-472] Remove some `has/getattr` in favor of `runtime_checkable` protocols

* [LABS-473] Fix `clvm_streamable` to prevent typing through wrapping

* [LABS-474] Make sure `crcat` is always bound

* [LABS-475] Fix basedpyright errors

* Comments by @cursor

* fix tests

* remove trade_store from mypy exclusions

* some easy wallet_rpc_api fixes

* Another easy fix for WalletRpcAPI

* [LABS-476] Make CATWallet subclassing less problematic

* Comment by @cursor

* fix tests

* Remove puzzle_for_pk methods as unused

* [LABS-477] Basedpyright fixes for wallet tests

* Comments by @cursor

* Bad merge

---------

+++ chia/_tests/wallet/cat_wallet/test_cat_wallet.py
@@ -1543,7 +1543,7 @@ async def test_unacknowledged_cat_table() -> None:
                 return CoinState(Coin(bytes32.zeros, bytes32.zeros, uint64(i)), None, None)
 
             await interested_store.add_unacknowledged_coin_state(asset_id(0), coin_state(0), None)
-            await interested_store.add_unacknowledged_coin_state(asset_id(1), coin_state(1), 100)
+            await interested_store.add_unacknowledged_coin_state(asset_id(1), coin_state(1), uint32(100))
             assert await interested_store.get_unacknowledged_states_for_asset_id(asset_id(0)) == [(coin_state(0), 0)]
             await interested_store.add_unacknowledged_coin_state(asset_id(0), coin_state(0), None)
             assert await interested_store.get_unacknowledged_states_for_asset_id(asset_id(0)) == [(coin_state(0), 0)]
